+++ cmd/main.go
@@ -2,15 +2,15 @@ package main
 
 import (
 	"context"
-	user "github.com/RyseUp/uniqora-nft-marketplace/backend/api/user/v1/v1connect"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/auth"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/config"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/models"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/mq"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/repositories"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/security"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/services"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/worker"
+	user "github.com/RyseUp/uniqora-nft-marketplace/api/user/v1/v1connect"
+	"github.com/RyseUp/uniqora-nft-marketplace/auth"
+	"github.com/RyseUp/uniqora-nft-marketplace/config"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/models"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/mq"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/repositories"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/security"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/services"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/worker"
 	"github.com/bufbuild/connect-go"
 	"go.uber.org/zap"
 	"log"
