+++ internal/mapper/user.go
@@ -1,8 +1,8 @@
 package mapper
 
 import (
-	v1 "github.com/RyseUp/uniqora-nft-marketplace/backend/api/user/v1"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/models"
+	v1 "github.com/RyseUp/uniqora-nft-marketplace/api/user/v1"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/models"
 )
 
 func ModelToProtoUserInfo(user *models.User) *v1.User {
