+++ internal/repositories/user_register.go
@@ -3,7 +3,7 @@ package repositories
 import (
 	"context"
 	"fmt"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/models"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/models"
 	"gorm.io/gorm"
 )
