+++ internal/services/user.go
@@ -8,15 +8,15 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	v1 "github.com/RyseUp/uniqora-nft-marketplace/backend/api/user/v1"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/api/user/v1/v1connect"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/auth"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/config"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/mapper"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/models"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/mq"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/repositories"
-	"github.com/RyseUp/uniqora-nft-marketplace/backend/internal/security"
+	v1 "github.com/RyseUp/uniqora-nft-marketplace/api/user/v1"
+	"github.com/RyseUp/uniqora-nft-marketplace/api/user/v1/v1connect"
+	"github.com/RyseUp/uniqora-nft-marketplace/auth"
+	"github.com/RyseUp/uniqora-nft-marketplace/config"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/mapper"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/models"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/mq"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/repositories"
+	"github.com/RyseUp/uniqora-nft-marketplace/internal/security"
 	"github.com/bufbuild/connect-go"
 	"github.com/google/uuid"
 	"go.uber.org/zap"
